Use correct worker type (lda, not lsi), and try a few more things
Was incorrectly using lsi workers instead of lda workers. Since the blocker appears to be getting the workers to talk to the nameserver, this wasn't causing problemsyet. Also, fixed the toy lda procedure to work when everything is run on one machine (which can be done by running the shell commands in terminals).